requirements.txt
# production
flask==1.0.2
Flask-Bootstrap4==4.0.2
blinker==1.4
flask-wtf==0.14.3
flask-login==0.4.1
wtforms-validators==1.0.0
firebase-admin==4.5.0
itsdangerous==1.1.0
Flask-Mail==0.9.1

# development
autopep8==1.5.4
Flask-Testing==0.8.0


